Controversial lifetime simplifications
I'm more than half way through the rust book and according to the lifetime elision rules described there
we could simplify some of the signatures.

While on one hand being explicit is good, there is an argument in favour of simplifying when possible:
adding unnecessary annotations (lifetime or type decls) is bound to draw some attention when the next
developer reads the code; they might pause to think why would that be necessary and it might thus cause distraction.