Use zip and find_map to unescape
This is possibly more elegant and also more robust as it panics if the escape sequence is not recognized.

That said, we apply our tests against the parser (I have a PR with a unit test for this unescape
but I realized it might be better to capture the test in the parser layer so it can be shared with
other implementations).